The_AO_ServerBackup Script
I created this script so experiment with the idea of creating a script that could easily be modified for use of backing up multiple directories. Simply changing the source and destination directory paths within the Variable Defintions section and the script takes care of the rest. The script concludes with a modifiable "cleanup" function that inspects the backup directories and removes the oldest files in place of the new one; and saving a user defined amount of backups at all times.